+++ src/web/server/main.js
@@ -1,11 +1,9 @@
 const MIME_TYPES={
 	"css": "text/css;charset=utf-8",
 	"deb": "application/octet-stream",
-	"dll": "application/octet-stream",
 	"gz": "application/gzip",
 	"html": "text/html;charset=utf-8",
 	"js": "application/javascript;charset=utf-8",
-	"so": "application/octet-stream",
 	"zip": "application/zip"
 };
 const REDIRECTS={